requirements.txt
numpy==1.21.2
scipy==1.7.2
scikit-learn==0.22
pytorch==1.10.0
gensim==4.1.2
nltk==3.6.5
tqdm==4.62.3